+++ lib/nestgen.dart
@@ -1,4 +1,4 @@
-/// arch_gen — a rich interactive CLI to scaffold Flutter projects with a
+/// nestgen — a rich interactive CLI to scaffold Flutter projects with a
 /// Clean Architecture or MVC folder structure.
 library;